pyproject.toml
[project]
name = "devlog-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "dotenv>=0.9.9",
    "fastapi[standard]>=0.135.3",
    "loadenv>=0.1.1",
    "pydantic>=2.12.5",
    "pyjwt>=2.12.1",
    "psycopg[binary]>=3.2.9",
    "sqlalchemy>=2.0.48",
    "pwdlib[argon2]>=0.3.0",
]
